Playing with experiment data.

+++ src/main/java/be/kuleuven/cs/flexsim/view/Grapher.java
@@ -116,7 +116,7 @@ public abstract class Grapher extends ApplicationFrame implements
 
         protected void record(Event e) {
             int t = e.getAttribute("time", Integer.class);
-            long y = e.getAttribute("totalTotalE", Long.class);
+            long y = e.getAttribute("totalTotalE", Integer.class);
             addRecord(STEPCONS, t, y);
         }
